refactor: split apollo-auth into core/doobie/http4s modules for Maven Central
Modules now follow the dependency boundaries of the code:

- apollo-core: type classes (HasId, HasEmail, HasPassword, Hashable) and
  the service traits (User/Session/Confirmation/Reset/MailService).
  Depends only on cats-core and cats-effect-std.
- apollo-doobie: Doobie/PostgreSQL implementations (DoobieUserService,
  DoobieSessionService, DoobieConfirmationService, DoobieResetService)
  plus the pg migration scripts. Fixes the uuid/varchar mismatch in
  updatePassword with an explicit ::uuid cast.
- apollo-http4s: AuthRoutes, the Apollo config aggregate, SessionAuth,
  CSRF/flash middleware, and the default Twirl templates. Trimmed to the
  deps the code actually uses (http4s-dsl/server/twirl, upickle) —
  ember, circe, and fs2 are no longer forced on consumers.

Other changes:
- MSSQL support removed (driver dep, SQLServerException handling,
  migration scripts); registration errors now map standard SQLState
  codes instead
- twirl contrib plugin pinned via $MILL_VERSION so it tracks the
  running Mill version
- artifact names derive from module names, so Scala-suffixed artifact
  ids (apollo-core_3, ...) are published as Maven Central requires
- README documents the module layout and the Sonatype Central Portal
  publishing workflow
