--- apollo-auth/src/db/mssql/001-create-confirmations.sql
@@ -1,19 +1,0 @@
-USE [webapp]
-GO
-
-SET ANSI_NULLS ON
-GO
-
-SET QUOTED_IDENTIFIER ON
-GO
-
-CREATE TABLE [dbo].[confirmations](
-	[user_id] [uniqueidentifier] NOT NULL,
-	[code] [nchar](32) NOT NULL,
-	[created_at] [datetimeoffset](7) NOT NULL
-) ON [PRIMARY]
-GO
-
-ALTER TABLE [dbo].[confirmations]  WITH CHECK ADD FOREIGN KEY([user_id])
-REFERENCES [dbo].[users] ([id])
-GO
