--- apollo-auth/src/me/joshuakfarrar/apollo/auth/ConfirmationService.scala
@@ -1,54 +1,0 @@
-package me.joshuakfarrar.apollo.auth
-
-import cats.data.{EitherT, OptionT}
-import cats.effect.Concurrent
-import cats.effect.std.Random
-import cats.implicits.*
-import doobie.*
-import doobie.implicits.*
-import me.joshuakfarrar.apollo.auth.util.generateAlphaNumericString
-
-trait ConfirmationService[F[_], U, I] {
-  def createConfirmation(user: U): EitherT[F, Throwable, String]
-
-  def confirmByCode(code: String): OptionT[F, Throwable]
-}
-
-object ConfirmationService {
-
-  def impl[F[_], U, I](
-      xa: Transactor[F]
-  )(using
-      R: Random[F],
-      C: Concurrent[F],
-      H: HasId[U, I]
-  ): ConfirmationService[F, U, I] =
-    new ConfirmationService[F, U, I] {
-
-      override def createConfirmation(
-          user: U
-      ): EitherT[F, Throwable, String] = for {
-        code <- EitherT.liftF(generateAlphaNumericString[F](32))
-        res <-
-          sql"""
-            INSERT INTO confirmations (user_id, code) 
-            VALUES (${H.id(user).toString}::uuid, $code)
-          """.update.run
-            .transact(xa)
-            .attemptT
-      } yield code
-
-      override def confirmByCode(code: String) =
-        (
-          sql"""
-            UPDATE users SET confirmed_at = NOW()
-            WHERE id = (SELECT user_id FROM confirmations WHERE code = $code)
-          """.update.run,
-          sql"DELETE FROM confirmations WHERE code = $code".update.run
-        ).mapN(_ + _)
-          .transact(xa)
-          .attemptT
-          .swap
-          .toOption
-    }
-}
