--- apollo-auth/src/me/joshuakfarrar/apollo/auth/HasEmail.scala
@@ -1,4 +1,0 @@
-package me.joshuakfarrar.apollo.auth
-
-trait HasEmail[U]:
-  def email(user: U): String
