--- apollo-auth/src/me/joshuakfarrar/apollo/auth/HasId.scala
@@ -1,4 +1,0 @@
-package me.joshuakfarrar.apollo.auth
-
-trait HasId[U, I]:
-  def id(user: U): I
