--- apollo-auth/src/me/joshuakfarrar/apollo/auth/HasPassword.scala
@@ -1,4 +1,0 @@
-package me.joshuakfarrar.apollo.auth
-
-trait HasPassword[U]:
-  def password(user: U): String
