--- apollo-auth/src/me/joshuakfarrar/apollo/auth/Hashable.scala
@@ -1,6 +1,0 @@
-package me.joshuakfarrar.apollo.auth
-
-trait Hashable[F[_], P] {
-  def hash(plain: String): F[P]
-  def verify(plain: String, hashed: P): F[Boolean]
-}
