+++ apollo-core/src/me/joshuakfarrar/apollo/core/HasPassword.scala
@@ -1,0 +1,4 @@
+package me.joshuakfarrar.apollo.core
+
+trait HasPassword[U]:
+  def password(user: U): String
