+++ apollo-core/src/me/joshuakfarrar/apollo/core/Hashable.scala
@@ -1,0 +1,6 @@
+package me.joshuakfarrar.apollo.core
+
+trait Hashable[F[_], P] {
+  def hash(plain: String): F[P]
+  def verify(plain: String, hashed: P): F[Boolean]
+}
