+++ build.mill
@@ -1,51 +1,63 @@
-//| mvnDeps: ["com.lihaoyi::mill-contrib-twirllib:1.0.4"]
+//| mvnDeps: ["com.lihaoyi::mill-contrib-twirllib:$MILL_VERSION"]
 package build
 
 import mill.*, scalalib.*, publish.*, mill.twirllib.*
 
 val Http4sVersion = "1.0.0-M44"
-//val Http4sVersion = "1.0-ecde17e-20250907T231823Z-SNAPSHOT"
-val CirceVersion = "0.14.14"
 val DoobieVersion = "1.0.0-RC10"
-val Fs2Version = "3.12.2"
+val CatsVersion = "2.13.0"
+val CatsEffectVersion = "3.6.3"
 
-object `apollo-auth` extends ScalaModule with TwirlModule with PublishModule {
+trait ApolloModule extends ScalaModule with PublishModule {
   def scalaVersion = "3.3.6"
-  def artifactId = "apollo-auth"
   def publishVersion = "0.1.0-SNAPSHOT"
+
+  def pomSettings = PomSettings(
+    description = "A Devise-like authentication library for Http4s",
+    organization = "me.joshuakfarrar",
+    url = "https://github.com/joshuakfarrar/apollo",
+    licenses = Seq(License.MIT),
+    versionControl = VersionControl.github("joshuakfarrar", "apollo"),
+    developers = Seq(
+      Developer(
+        "joshuakfarrar",
+        "Joshua K. Farrar",
+        "https://github.com/joshuakfarrar"
+      )
+    )
+  )
+}
+
+object `apollo-core` extends ApolloModule {
+  def mvnDeps = Seq(
+    mvn"org.typelevel::cats-core:$CatsVersion",
+    mvn"org.typelevel::cats-effect-std:$CatsEffectVersion"
+  )
+}
+
+object `apollo-doobie` extends ApolloModule {
+  def moduleDeps = Seq(`apollo-core`)
+
+  def mvnDeps = Seq(
+    mvn"org.tpolecat::doobie-core:$DoobieVersion"
+  )
+}
+
+object `apollo-http4s` extends ApolloModule with TwirlModule {
   def twirlVersion = "2.0.9"
+
+  def moduleDeps = Seq(`apollo-core`)
 
   override def generatedSources = Task {
     Seq(compileTwirl().classes)
   }
 
-  def pomSettings = PomSettings(
-    organization = "me.joshuakfarrar",
-    description = "An authentication solution for Http4s",
-    url = "https://github.com/joshuakfarrar/apollo",
-    licenses = Seq(License.MIT),
-    versionControl = VersionControl(browsableRepository = Some("https://github.com/joshuakfarrar/apollo")),
-    developers = Seq(Developer(
-      id = "joshuakfarrar",
-      name = "Joshua K. Farrar",
-      url = "https://github.com/joshuakfarrar"
-    ))
-  )
-
   def mvnDeps = Seq(
-    mvn"com.microsoft.sqlserver:mssql-jdbc:13.2.0.jre11",
-    mvn"org.http4s::http4s-ember-server:$Http4sVersion",
-    mvn"org.http4s::http4s-ember-client:$Http4sVersion",
     mvn"org.http4s::http4s-dsl:$Http4sVersion",
+    mvn"org.http4s::http4s-server:$Http4sVersion",
     mvn"org.http4s::http4s-crypto:0.2.5-RC1",
-    mvn"org.typelevel::cats-core:2.13.0",
-    mvn"org.typelevel::cats-effect:3.6.3",
-    mvn"org.tpolecat::doobie-core:$DoobieVersion",
-    mvn"io.circe::circe-generic:$CirceVersion",
-    mvn"org.http4s::http4s-circe:$Http4sVersion",
+    // no M44 build of http4s-twirl exists; M38 is the latest
     mvn"org.http4s::http4s-twirl:1.0.0-M38",
-    mvn"co.fs2::fs2-core:$Fs2Version",
-    mvn"co.fs2::fs2-io:$Fs2Version",
-    mvn"com.lihaoyi::upickle:4.1.0",
+    mvn"com.lihaoyi::upickle:4.1.0"
   )
 }
